+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## New
 
++ `chart_spreads()` to generate specific contract spreads across years e.g. ULSD March/April. Requires Morninstar credentials.
+
 + Morningstar feeds: 
 
     + CME_Comex_FuturesSettlement_EOD and CME_Comex_FuturesSettlement_EOD_continuous.
@@ -13,6 +15,7 @@
 
 + `eia2tidy()` amended for quarterly and hourly data.
 + Morningstar tickers for `getPrice()` and `getPrices()` functions.
++ `promptBeta()` chart moved to `plotly`.
 
 ## Removed